+++ jannovar-core/src/test/java/de/charite/compbio/jannovar/impl/util/TranslatorTest.java
@@ -24,8 +24,10 @@ public class TranslatorTest {
 	 */
 	@Test
 	public void testTranslateDna_tooShort() throws AnnotationException {
-		Assertions.assertEquals("", translator.translateDNA("A"));
-		Assertions.assertEquals("", translator.translateDNA("AC"));
+		Assertions.assertEquals("", translator.translateDNA("A", true));
+		Assertions.assertEquals("", translator.translateDNA("AC", true));
+		Assertions.assertEquals("", translator.translateDNA("A", false));
+		Assertions.assertEquals("", translator.translateDNA("AC", false));
 	}
 
 	/**
@@ -33,7 +35,8 @@ public class TranslatorTest {
 	 */
 	@Test
 	public void testTranslateDna_short() throws AnnotationException {
-		Assertions.assertEquals("T", translator.translateDNA("ACT"));
+		Assertions.assertEquals("T", translator.translateDNA("ACT", true));
+		Assertions.assertEquals("T", translator.translateDNA("ACT", false));
 	}
 
 	/**
@@ -41,7 +44,9 @@ public class TranslatorTest {
 	 */
 	@Test
 	public void testTranslateDna_longer() throws AnnotationException {
-		Assertions.assertEquals("M*S", translator.translateDNA("ATGTAGAGT"));
+		Assertions.assertEquals("M*S", translator.translateDNA("ATGTAGAGT", true));
+		Assertions.assertEquals("**S", translator.translateDNA("TGATAGAGT", true));
+		Assertions.assertEquals("W*S", translator.translateDNA("TGATAGAGT", false));
 	}
 
 	/**
@@ -49,6 +54,40 @@ public class TranslatorTest {
 	 */
 	@Test
 	public void testTranslateDna_tooLonger() throws AnnotationException {
-		Assertions.assertEquals("T", translator.translateDNA("ACTG"));
+		Assertions.assertEquals("T", translator.translateDNA("ACTG", true));
+		Assertions.assertEquals("*", translator.translateDNA("TGAT", true));
+		Assertions.assertEquals("W", translator.translateDNA("TGAT", false));
+	}
+
+	/**
+	 * Test for mitochondrial code exceptions (1 character)
+	 */
+	@Test
+	public void testTranslateChrMtExceptions_short() throws AnnotationException {
+		Assertions.assertEquals("R", translator.translateDNA("AGA", true));
+		Assertions.assertEquals("R", translator.translateDNA("AGG", true));
+		Assertions.assertEquals("I", translator.translateDNA("ATA", true));
+		Assertions.assertEquals("*", translator.translateDNA("TGA", true));
+
+		Assertions.assertEquals("*", translator.translateDNA("AGA", false));
+		Assertions.assertEquals("*", translator.translateDNA("AGG", false));
+		Assertions.assertEquals("M", translator.translateDNA("ATA", false));
+		Assertions.assertEquals("W", translator.translateDNA("TGA", false));
+	}
+
+	/**
+	 * Test for mitochondrial code exceptions (3 characters)
+	 */
+	@Test
+	public void testTranslateChrMtExceptions_long() throws AnnotationException {
+		Assertions.assertEquals("Arg", translator.translateDNA3("AGA", true));
+		Assertions.assertEquals("Arg", translator.translateDNA3("AGG", true));
+		Assertions.assertEquals("Ile", translator.translateDNA3("ATA", true));
+		Assertions.assertEquals("*", translator.translateDNA3("TGA", true));
+
+		Assertions.assertEquals("*", translator.translateDNA3("AGA", false));
+		Assertions.assertEquals("*", translator.translateDNA3("AGG", false));
+		Assertions.assertEquals("Met", translator.translateDNA3("ATA", false));
+		Assertions.assertEquals("Trp", translator.translateDNA3("TGA", false));
 	}
 }