+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Jannovar Changelog
 
+## HEAD (unreleased)
+
+### overall
+
+- moved away from split between `master` and `develop` branch to using `main` branch
+- adding support for mitochondrial genetic code (#501)
+
 ## v0.38
 
 ### overall